requirements.txt
requests>=2.31.0
pandas>=2.0.0
jieba>=0.42.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
openai>=1.0.0
google-genai>=1.0.0
pypinyin>=0.50.0

# GUI
customtkinter>=5.2.2
